==> jupyter_nooteboks/01_Etl_games.ipynb <==
importaciones

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn as sk
import fastapi as fp
import ast

extraccion de la data y primer vistazo


archivo = '../archivos_crudo/output_steam_games.json'

df_games = pd.read_json(archivo, lines=True)
df_games

aca busco ver las columnas y la cantidad de datos faltantes


df_games.info()

df_games.dtypes

aca busco ver la cantidad de nulos de cada columna



df_games.isnull().sum()

realizo una prueba para ver como quedarian los datos si se quitaran todos los datos faltantes


df_1 = df_games.dropna()
df_1

aca quite las primeras 88310 filas ya que no contenian ningun tipo de dato


df_games = df_games.iloc[88310:, :]
df_games

busco ver los nulos de vuelta para tener una vision mas acertada

df_games.isnull().sum()

procedo a dropear columnas de baja utilidad para mi

"""
las siguientes columnas se eliminaron por lo siguiente:
early_access se elimino por la falta de utilidad para el modelo o como tipo de informacion
specs solo contenia datos que no eran utiles para la clasificacion para el modelo ML
title se elimina ya que la columna app_name ya tiene los mismos datos y mas completos
publisher se elimina por cantidad de datos faltantes
"""

df_games.drop(['early_access', 'specs', 'title','publisher'], axis=1, inplace=True)

busco las filas con valores faltantes en la columna Id y los elimino

# Encuentra la fila con valores faltantes en la columna ID
dato_faltante_ID = df_games['id'].isna()

# me devuelve esas filas
df_games[dato_faltante_ID]




#elimino esas filas
df_games = df_games.drop([88384,119271])

df_exploded = df_games.explode('genres')
valores_unicos = df_exploded['genres'].unique()
print(valores_unicos)

ahora procedo a tratar los datos nullos que me quedan

# ya que la columna tags y genres tienen valores parecidos con algunas diferencias
df_games['genres'] = df_games['genres'].fillna(df_games['tags'])
df_games


verifico el el estado de los nulos

df_games.isnull().sum()

se dropea la columna 'tags' ya que ya obtuve los datos utiles de la misma

df_games.drop(['tags'], axis=1, inplace = True)


los siguientes pasos son para buscar generalizar los generos y crear cierto orden

df_exploded = df_games.explode('genres')
valores_unicos = df_exploded['genres'].unique()

print(valores_unicos)

# los generos se sacaron de la lista de generos antes de utilizar la columna tags para completar info. sacando: free to play, early 
# access
generos_juegos = ["Action", "Casual", "Indie", "Simulation", "Strategy", "RPG", "Sports", "Adventure", "Racing", "Massively Multiplayer", "Animation & Modeling", "Video Production", "Utilities", "Web Publishing", "Education", "Software Training", "Design & Illustration", "Audio Production", "Photo Editing", "Accounting"]

# Elimina los elementos duplicados de la lista
generos_juegos = list(set(generos_juegos))

print(generos_juegos)

# Convertir la lista en un conjunto
lista_comparacion = set(generos_juegos)

# Filtrar el DataFrame por las filas que coinciden con el conjunto
df_games = df_exploded.loc[df_exploded['genres'].isin(lista_comparacion)]

df_games


ahora pasaria a la normalizacion de las demas columnas para que no tengan nulos/vacios/etc, para el correcto funcionamiento del modelo

correccion = "Valor no disponible"

df_games['app_name'].fillna(correccion, inplace=True)
df_games['release_date'].fillna(correccion, inplace=True)
df_games['developer'].fillna(correccion, inplace=True)


df_games.isnull().sum()

df_games

normalizando columna 'price'

df_games['price'] = df_games['price'].replace('Free To Play', 0)
df_games['price'] = df_games['price'].replace('Free', 0)
df_games['price'] = df_games['price'].replace('Free to Play', 0)
df_games['price'] = df_games['price'].replace('Free Demo', 0)
df_games['price'] = df_games['price'].replace('Play for Free!', 0)
df_games['price'] = df_games['price'].replace('Install Now', 0)
df_games['price'] = df_games['price'].replace('Play WARMACHINE: Tactics Demo', 0)
df_games['price'] = df_games['price'].replace('Free Mod', 0)
df_games['price'] = df_games['price'].replace('Install Theme', 0)
df_games['price'] = df_games['price'].replace('Third-party', 0)
df_games['price'] = df_games['price'].replace('Play Now', 0)
df_games['price'] = df_games['price'].replace('Free HITMAN™ Holiday Pack', 0)
df_games['price'] = df_games['price'].replace('Play the Demo', 0)
df_games['price'] = df_games['price'].replace('Starting at $449.00', 449)
df_games['price'] = df_games['price'].replace('Free to Try', 449)
df_games['price'] = df_games['price'].replace('Free to Use', 449)
df_games['price'] = df_games['price'].fillna(0)

df_games

#df_games['price'] = df_games['price'].astype(tipo_dato)


df_games['price'] = df_games['price'].astype(float)


df_games

df_games.isnull().sum()

con todas las columnas normalizadas concluyo con el etl del archivo

# el siguiente paso es guardar la data limpia para posterior uso en el modelo ML
df_games.to_parquet('games.parquet')
