==> jupyter_nooteboks/01_Etl_items.ipynb <==
importaciones

import pandas as pd
import numpy as np
import json
import ast

extracion de la data y primeros vistazos

# Ruta al dataset australian_user_reviews
ruta_items = '../archivos_crudo/australian_users_items.json'

# Se lee de cada línea del dataset
filas_items = []
with open(ruta_items) as f:
    for line in f.readlines():
        filas_items.append(ast.literal_eval(line))

# Se convierte en dataframe
df_items = pd.DataFrame(filas_items)
df_items
#df_items = pd.read_parquet('archivos_en_parquet/australian_users_items.parquet')


realizo un escaneo rapido

df_items.info()

df_items.isnull().sum()

al parecer encontramos otra vez una columna con anidaciones por lo que procederemos a normalizarla

df_items

# se normaliza la columna 'items'
df_items2 = pd.json_normalize(filas_items, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
df_items2

# buscando si hay duplicados

print(df_items2.duplicated().sum())


# se encontraron 59104 duplicados por los que se eliminaran
# creo un dataframe con los duplicados
duplicate_ids = df_items2.duplicated()
df_items2[duplicate_ids]

# elimino los id´s duplicados
df_items2 = df_items2[~duplicate_ids] 
df_items2

# se elimina la columna playtime_2weeks ya que no es de utilidad

df_items2 = df_items2.drop('playtime_2weeks', axis=1)
df_items2

df_items2.to_parquet('items.parquet')
