==> jupyter_nooteboks/01_Etl_review.ipynb <==
importaciones

import pandas as pd
from pandas import json_normalize
import numpy as np
import json
import ast
import re
from textblob import TextBlob

importacion de la data y primeros vistazos

ruta_review = '../archivos_crudo/australian_user_reviews.json'

# Se lee de cada línea del dataset
filas_review = []
with open(ruta_review) as f:
    for line in f.readlines():
        filas_review.append(ast.literal_eval(line))

# Se convierte en dataframe
df_reviews = pd.DataFrame(filas_review)
df_reviews.to_parquet

# le echo un primer vistazo a la data 
df_reviews

df_reviews.info()

df_reviews.isnull().sum()

empiezo eliminando los id duplicados 

# verifico si hay o no duplicados por cada 'user_id' y cuantos
print(df_reviews['user_id'].duplicated().sum())


# creo un dataframe con los duplicados
duplicate_ids = df_reviews['user_id'].duplicated()
df_reviews[duplicate_ids]

# elimino los id´s duplicados
df_reviews = df_reviews[~duplicate_ids] 
df_reviews

empiezo con el desanidado de la columna reviews y su normalizacion

# Se transforma a columnas cada elemento de las listas
df_reviews2 = pd.json_normalize(df_reviews['reviews'])
df_reviews2.head()

# Se agrega el 'user_id' y 'user_url' a las columnas separadas 
df_reviews2 = pd.concat([df_reviews[['user_id', 'user_url']], df_reviews2], axis=1)
df_reviews2

# Se utiliza pd.melt para transformar las columnas en filas conservando el 'user_id' y 'user_url'
df_reviews2 = pd.melt(df_reviews2, id_vars=['user_id', 'user_url'], 
                       value_vars=list(range(9)),
                       value_name='reviews')
df_reviews2

# Se eliminan las filas con valor None
df_reviews2 = df_reviews2.dropna()

# Se separan por columnas cada una de las claves de 'reviews'
df_reviews = df_reviews2['reviews'].apply(pd.Series, dtype='object')
df_reviews = df_reviews.add_prefix('reviews_')
df_reviews.head()

# Se une con el 'user_id' y 'user_url' 
df_reviews = pd.concat([df_reviews2[['user_id', 'user_url']], df_reviews], axis=1)
df_reviews.head()

df_reviews

df_reviews['reviews_last_edited'].isnull().sum()

df_reviews['reviews_posted'].isnull().sum()

# Se eliminan las columnas 'reviews_funny' y 'reviews_last_edited' 
#ya que considero que ya no son de utilidad
df_reviews = df_reviews.drop(columns=['reviews_funny', 'reviews_last_edited'])
df_reviews

df_reviews.info()

df_reviews.isnull().sum()

yo considero que la data ya fue tratada y limpiada. por lo que finalizo el Etl del mismo y lo guardo en archivos limpios


df_reviews.to_parquet('reviews.parquet')
