==> jupyter_nooteboks/02_Eda_all.ipynb <==
EDA


importaciones

import pandas as pd 
import matplotlib.pyplot as plt 
import seaborn as sns 
import pyarrow as pa 
import pyarrow.parquet as pq 
import warnings

importamos los datos y empezamos con el eda


df_games = pd.read_parquet('../archivos_limpios/games.parquet')

df_items = pd.read_parquet('../archivos_limpios/items.parquet')

df_reviews = pd.read_parquet('../archivos_limpios/reviews.parquet')

empezamos por el de games

df_games.info()

# vemos la cantidad de nulos
df_games.isnull().sum()

df_games.head()

# vemos la cantidad de juegos unicos
games_unique = len(df_games['id'].unique()) #con esto obtenemos la cantidad de juegos unicos
print('se obsevan: ',games_unique, ' juegos unicos')

# aca vemos la cantidad de juegos por desarrollador
developers = df_games['developer'].value_counts() #obtenemos la cantidad de jeugos por desarrollador
developers

# vemos cuantos juegos hay por su año de lanzamiento
games_years = df_games['release_date'].value_counts()
games_years

seguimos con el archivo de items

df_items.info()

# vemos los nulos en los datos
df_items.isnull().sum()

#vemos la cantidad de juegos en el dataframe
games_unique = len(df_items['item_name'].unique())
print('se obsevan: ',games_unique, ' juegos unicos')

#vemos la cantidad de juegos por nombre
games = df_items['item_name'].value_counts()
games

seguimos ahora con el dataset de reviews

#vemos la informacion de las columnas del dataframe
df_reviews.info()

# vemos la cantidad de nulos 
df_reviews.isnull().sum() 

#Vemos la cantidad de usuarios que dejaron reviews
user_unique= len(df_reviews['user_id'].unique())
print('se observan:',user_unique,'usuarios distintos que hallan escrito reviews')

#vemos la cantidad de recomendaciones hechas
write_reviews = df_reviews['reviews_recommend'].value_counts()
write_reviews
#se observan 50778 reviews escritas

#vemos el conteo del analisis de sentimiento para cada caso
sentiment_analysis = df_reviews['sentiment_analysis'].value_counts()
sentiment_analysis
# esto señala que hicieron: (2) 17100 reviews positivas
#                           (1) 35271 review neutras   
#                           (0) 5026 reviews negativas
